requirements.txt
pykeepass==3.0.2
python_libxdo==0.1.2a1
Click==7.0
